chore: adds transfer and incoming_transactions to rpc client

+++ webapp/Wallet/DataTransfer/AddressDto.cs
@@ -17,7 +17,7 @@ namespace xmr_tutorials.Wallet.DataTransfer
         public string AddressAddress { get; set; }
 
         [JsonProperty("address_index")]
-        public long AddressIndex { get; set; }
+        public uint AddressIndex { get; set; }
 
         [JsonProperty("label")]
         public string Label { get; set; }
